Implement Subordinate (inverse ordinal sum) calculation for short games and stoppers (#35)
Adds a method CanonicalStopper.Subordinate(B) that calculates the unique H for which G = B:H (or raises an error if no such H exists). B is permitted to be any short game, not necessarily canonical.

There is a separate implementation for CanonicalShortGames that uses a recursive calculation and is more efficient.