Alters the behavior of CanonicalStopper.Variety to reflect that the Stability Conjecture is false. (#47)
Previously, G.Variety displayed the following message in cases where G is unstable: "Congratulations! You've found a counterexample to the Stability Conjecture."

In 2022 Ian Grenville found such a counterexample, so this commit changes the message to: "That game is not stable."

It also introduces G.IsStable with the obvious meaning.