Transfinite Nim Arithmetic (#44)
Implements the theory of transfinite nim arithmetic as described by Conway and Lenstra.

This commit introduces a new CGScript class, `TransfiniteNimber`, together with methods `GeneralizedOrdinal.NimSum` and `GeneralizedOrdinal.NimProduct` that implement the relevant operations for ordinals below omega^omega^omega.

The fundamental constants for nim multiplication to a given ordinal ("Q-Set" and "excess"; cf. Lenstra 1977) are computationally expensive to produce, so they are hard-coded in the `NimProduct` implementation up to `p = 257`. The tool `org.cgsuite.core.impartial.arithmetic.NimFieldCalculator` can be used to regenerate them or to extend the list.

Also in this commit are a few related methods, `Integer.NimInverse` and `Integer.NimDiv`.